Troca o nome entre txt limpo e puro
