+++ Capture_TXT/DataMiner/CleanTxt/Commitments.py
