+++ Capture_TXT/DataMiner/CleanTxt/Lawsuit.py
