+++ Capture_TXT/DataMiner/CleanTxt/Pendence.py
