+++ Capture_TXT/DataMiner/CleanTxt/Protest.py
