+++ Capture_TXT/DataMiner/CleanTxt/RegistryConsults.py
