+++ Capture_TXT/DataMiner/CleanTxt/Relationship.py
