+++ Capture_TXT/DataMiner/PureTxt/R301.py
