+++ Capture_TXT/DataMiner/PureTxt/R302.py
