+++ Capture_TXT/DataMiner/PureTxt/R451Evolucao.py
