--- Capture_TXT/DataMiner/PureTxt/__init__.py
@@ -1,10 +1,0 @@
-from .BuisinessReferences import *
-from .RegistryConsults import *
-from .PaymentsHistory import *
-from .Relationship import *
-from .OverdueDebt import *
-from .Commitments import *
-from .Bankruptcy import *
-from .Pendence import *
-from .Lawsuit import *
-from .Protest import *
